Adding saved_objects_page in OSS (#69900)
* add savedObjects own PO

* fix usage

* simplify functions

* fix test

* fix title parsing

* add missing await

* improve parsing

* wait for table is loaded
