+++ test/functional/apps/management/_import_objects.js
@@ -24,7 +24,7 @@ import { indexBy } from 'lodash';
 export default function ({ getService, getPageObjects }) {
   const kibanaServer = getService('kibanaServer');
   const esArchiver = getService('esArchiver');
-  const PageObjects = getPageObjects(['common', 'settings', 'header']);
+  const PageObjects = getPageObjects(['common', 'settings', 'header', 'savedObjects']);
   const testSubjects = getService('testSubjects');
   const log = getService('log');
 
@@ -43,22 +43,19 @@ export default function ({ getService, getPageObjects }) {
       });
 
       it('should import saved objects', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects.ndjson')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
         // get all the elements in the table, and index them by the 'title' visible text field
-        const elements = indexBy(
-          await PageObjects.settings.getSavedObjectElementsInTable(),
-          'title'
-        );
+        const elements = indexBy(await PageObjects.savedObjects.getElementsInTable(), 'title');
         log.debug("check that 'Log Agents' is in table as a visualization");
         expect(elements['Log Agents'].objectType).to.eql('visualization');
 
         await elements['logstash-*'].relationshipsElement.click();
-        const flyout = indexBy(await PageObjects.settings.getRelationshipFlyout(), 'title');
+        const flyout = indexBy(await PageObjects.savedObjects.getRelationshipFlyout(), 'title');
         log.debug(
           "check that 'Shared-Item Visualization AreaChart' shows 'logstash-*' as it's Parent"
         );
@@ -68,18 +65,18 @@ export default function ({ getService, getPageObjects }) {
       });
 
       it('should provide dialog to allow the importing of saved objects with index pattern conflicts', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_conflicts.ndjson')
         );
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern(
           'd1e4c910-a2e6-11e7-bb30-233be9be6a15',
           'logstash-*'
         );
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
         await PageObjects.header.waitUntilLoadingHasFinished();
-        await PageObjects.settings.clickImportDone();
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        await PageObjects.savedObjects.clickImportDone();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object with index pattern conflict');
         expect(isSavedObjectImported).to.be(true);
       });
@@ -87,14 +84,14 @@ export default function ({ getService, getPageObjects }) {
       it('should allow the user to override duplicate saved objects', async function () {
         // This data has already been loaded by the "visualize" esArchive. We'll load it again
         // so that we can override the existing visualization.
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_exists.ndjson'),
           false
         );
 
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern('logstash-*', 'logstash-*');
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
 
         // Override the visualization.
         await PageObjects.common.clickConfirmOnModal();
@@ -106,14 +103,14 @@ export default function ({ getService, getPageObjects }) {
       it('should allow the user to cancel overriding duplicate saved objects', async function () {
         // This data has already been loaded by the "visualize" esArchive. We'll load it again
         // so that we can be prompted to override the existing visualization.
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_exists.ndjson'),
           false
         );
 
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern('logstash-*', 'logstash-*');
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
 
         // *Don't* override the visualization.
         await PageObjects.common.clickCancelOnModal();
@@ -123,86 +120,80 @@ export default function ({ getService, getPageObjects }) {
       });
 
       it('should import saved objects linked to saved searches', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_saved_search.ndjson')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_connected_to_saved_search.ndjson')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(true);
       });
 
       it('should not import saved objects linked to saved searches when saved search does not exist', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_connected_to_saved_search.ndjson')
         );
-        await PageObjects.settings.checkNoneImported();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkNoneImported();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(false);
       });
 
       it('should not import saved objects linked to saved searches when saved search index pattern does not exist', async function () {
-        const elements = indexBy(
-          await PageObjects.settings.getSavedObjectElementsInTable(),
-          'title'
-        );
+        const elements = indexBy(await PageObjects.savedObjects.getElementsInTable(), 'title');
         await elements['logstash-*'].checkbox.click();
-        await PageObjects.settings.clickSavedObjectsDelete();
+        await PageObjects.savedObjects.clickDelete();
 
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_with_saved_search.ndjson')
         );
         // Wait for all the saves to happen
-        await PageObjects.settings.checkImportConflictsWarning();
-        await PageObjects.settings.clickConfirmChanges();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
+        await PageObjects.savedObjects.clickConfirmChanges();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(false);
       });
 
       it('should import saved objects with index patterns when index patterns already exists', async () => {
         // First, import the objects
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_with_index_patterns.ndjson')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object imported with index pattern');
         expect(isSavedObjectImported).to.be(true);
       });
 
       it('should import saved objects with index patterns when index patterns does not exists', async () => {
         // First, we need to delete the index pattern
-        const elements = indexBy(
-          await PageObjects.settings.getSavedObjectElementsInTable(),
-          'title'
-        );
+        const elements = indexBy(await PageObjects.savedObjects.getElementsInTable(), 'title');
         await elements['logstash-*'].checkbox.click();
-        await PageObjects.settings.clickSavedObjectsDelete();
+        await PageObjects.savedObjects.clickDelete();
 
         // Then, import the objects
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_with_index_patterns.ndjson')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object imported with index pattern');
         expect(isSavedObjectImported).to.be(true);
       });
@@ -222,30 +213,30 @@ export default function ({ getService, getPageObjects }) {
       });
 
       it('should import saved objects', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects.json')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('Log Agents');
         expect(isSavedObjectImported).to.be(true);
       });
 
       it('should provide dialog to allow the importing of saved objects with index pattern conflicts', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects-conflicts.json')
         );
-        await PageObjects.settings.checkImportLegacyWarning();
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportLegacyWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern(
           'd1e4c910-a2e6-11e7-bb30-233be9be6a15',
           'logstash-*'
         );
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
         await PageObjects.header.waitUntilLoadingHasFinished();
-        await PageObjects.settings.clickImportDone();
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        await PageObjects.savedObjects.clickImportDone();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object with index pattern conflict');
         expect(isSavedObjectImported).to.be(true);
       });
@@ -253,15 +244,15 @@ export default function ({ getService, getPageObjects }) {
       it('should allow the user to override duplicate saved objects', async function () {
         // This data has already been loaded by the "visualize" esArchive. We'll load it again
         // so that we can override the existing visualization.
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_exists.json'),
           false
         );
 
-        await PageObjects.settings.checkImportLegacyWarning();
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportLegacyWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern('logstash-*', 'logstash-*');
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
 
         // Override the visualization.
         await PageObjects.common.clickConfirmOnModal();
@@ -273,15 +264,15 @@ export default function ({ getService, getPageObjects }) {
       it('should allow the user to cancel overriding duplicate saved objects', async function () {
         // This data has already been loaded by the "visualize" esArchive. We'll load it again
         // so that we can be prompted to override the existing visualization.
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_exists.json'),
           false
         );
 
-        await PageObjects.settings.checkImportLegacyWarning();
-        await PageObjects.settings.checkImportConflictsWarning();
+        await PageObjects.savedObjects.checkImportLegacyWarning();
+        await PageObjects.savedObjects.checkImportConflictsWarning();
         await PageObjects.settings.associateIndexPattern('logstash-*', 'logstash-*');
-        await PageObjects.settings.clickConfirmChanges();
+        await PageObjects.savedObjects.clickConfirmChanges();
 
         // *Don't* override the visualization.
         await PageObjects.common.clickCancelOnModal();
@@ -291,95 +282,89 @@ export default function ({ getService, getPageObjects }) {
       });
 
       it('should import saved objects linked to saved searches', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_saved_search.json')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_connected_to_saved_search.json')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(true);
       });
 
       it('should not import saved objects linked to saved searches when saved search does not exist', async function () {
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_connected_to_saved_search.json')
         );
-        await PageObjects.settings.checkImportFailedWarning();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportFailedWarning();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(false);
       });
 
       it('should not import saved objects linked to saved searches when saved search index pattern does not exist', async function () {
         // First, import the saved search
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_saved_search.json')
         );
         // Wait for all the saves to happen
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
         // Second, we need to delete the index pattern
-        const elements = indexBy(
-          await PageObjects.settings.getSavedObjectElementsInTable(),
-          'title'
-        );
+        const elements = indexBy(await PageObjects.savedObjects.getElementsInTable(), 'title');
         await elements['logstash-*'].checkbox.click();
-        await PageObjects.settings.clickSavedObjectsDelete();
+        await PageObjects.savedObjects.clickDelete();
 
         // Last, import a saved object connected to the saved search
         // This should NOT show the conflicts
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_connected_to_saved_search.json')
         );
         // Wait for all the saves to happen
-        await PageObjects.settings.checkNoneImported();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkNoneImported();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object connected to saved search');
         expect(isSavedObjectImported).to.be(false);
       });
 
       it('should import saved objects with index patterns when index patterns already exists', async () => {
         // First, import the objects
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_with_index_patterns.json')
         );
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object imported with index pattern');
         expect(isSavedObjectImported).to.be(true);
       });
 
       it('should import saved objects with index patterns when index patterns does not exists', async () => {
         // First, we need to delete the index pattern
-        const elements = indexBy(
-          await PageObjects.settings.getSavedObjectElementsInTable(),
-          'title'
-        );
+        const elements = indexBy(await PageObjects.savedObjects.getElementsInTable(), 'title');
         await elements['logstash-*'].checkbox.click();
-        await PageObjects.settings.clickSavedObjectsDelete();
+        await PageObjects.savedObjects.clickDelete();
 
         // Then, import the objects
-        await PageObjects.settings.importFile(
+        await PageObjects.savedObjects.importFile(
           path.join(__dirname, 'exports', '_import_objects_with_index_patterns.json')
         );
-        await PageObjects.settings.checkImportSucceeded();
-        await PageObjects.settings.clickImportDone();
+        await PageObjects.savedObjects.checkImportSucceeded();
+        await PageObjects.savedObjects.clickImportDone();
 
-        const objects = await PageObjects.settings.getSavedObjectsInTable();
+        const objects = await PageObjects.savedObjects.getRowTitles();
         const isSavedObjectImported = objects.includes('saved object imported with index pattern');
         expect(isSavedObjectImported).to.be(true);
       });
