+++ test/functional/apps/management/_mgmt_import_saved_objects.js
@@ -22,7 +22,7 @@ import path from 'path';
 
 export default function ({ getService, getPageObjects }) {
   const esArchiver = getService('esArchiver');
-  const PageObjects = getPageObjects(['common', 'settings', 'header']);
+  const PageObjects = getPageObjects(['common', 'settings', 'header', 'savedObjects']);
 
   //in 6.4.0 bug the Saved Search conflict would be resolved and get imported but the visualization
   //that referenced the saved search was not imported.( https://github.com/elastic/kibana/issues/22238)
@@ -40,19 +40,19 @@ export default function ({ getService, getPageObjects }) {
 
     it('should import saved objects mgmt', async function () {
       await PageObjects.settings.clickKibanaSavedObjects();
-      await PageObjects.settings.importFile(
+      await PageObjects.savedObjects.importFile(
         path.join(__dirname, 'exports', 'mgmt_import_objects.json')
       );
       await PageObjects.settings.associateIndexPattern(
         '4c3f3c30-ac94-11e8-a651-614b2788174a',
         'logstash-*'
       );
-      await PageObjects.settings.clickConfirmChanges();
-      await PageObjects.settings.clickImportDone();
-      await PageObjects.settings.waitUntilSavedObjectsTableIsNotLoading();
+      await PageObjects.savedObjects.clickConfirmChanges();
+      await PageObjects.savedObjects.clickImportDone();
+      await PageObjects.savedObjects.waitTableIsLoaded();
 
       //instead of asserting on count- am asserting on the titles- which is more accurate than count.
-      const objects = await PageObjects.settings.getSavedObjectsInTable();
+      const objects = await PageObjects.savedObjects.getRowTitles();
       expect(objects.includes('mysavedsearch')).to.be(true);
       expect(objects.includes('mysavedviz')).to.be(true);
     });
